Revert web changes
