+++ app/components/Pledge/index.js
@@ -7,6 +7,7 @@ import { getLocalRoute } from '../../actions/apiRouting';
 import { getDocumentTitle } from '../../helpers/utils';
 import { delimitNumbers } from '../../utils/utils';
 import Modal from 'react-modal';
+import { fetchItem } from './../../stores/localStorage';
 import {
   pledgeFormSchema,
   pledgeSchemaOptions
@@ -21,7 +22,6 @@ import {
 import { AllPledges, IncreasePledge, MakePledge, UpdatePledge } from './modals';
 let TCombForm = t.form.Form;
 import Fade from 'react-reveal/Fade';
-import LoadingIndicator from '../Common/LoadingIndicator';
 
 const formLayout = locals => {
   return (
@@ -55,16 +55,47 @@ export default class Pledge extends Component {
       loggedIn: false,
       myPledge: {},
       updatingTreeCount: '', // used while increasing tree count
-      pledgeButtonDisabled: false, // To disable button until page reloads
-      loading: true
+      pledgeButtonDisabled: false // To disable button until page reloads
     };
   }
 
-  componentDidUpdate() {
-    if (this.props.pledges && this.props.pledges.image && this.state.loading) {
+  componentDidMount() {
+    if (this.props.currentUserProfile) {
       this.setState({
-        loading: false
+        loggedIn: true
       });
+    } else {
+      fetchItem('pledgedEvent')
+        .then(data => {
+          if (typeof data !== 'undefined' && data.length > 0) {
+            let stringPledges = JSON.parse(data);
+            stringPledges = stringPledges.toString();
+            this.props.fetchPublicPledgesAction(stringPledges);
+          }
+        })
+        .catch(error => console.log(error));
+    }
+    this.getMyPledge();
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.pledges !== this.props.pledges) {
+      if (this.props.currentUserProfile) {
+        this.setState({
+          loggedIn: true
+        });
+      } else {
+        fetchItem('pledgedEvent')
+          .then(data => {
+            if (typeof data !== 'undefined' && data.length > 0) {
+              let stringPledges = JSON.parse(data);
+              stringPledges = stringPledges.toString();
+              this.props.fetchPublicPledgesAction(stringPledges);
+            }
+          })
+          .catch(error => console.log(error));
+      }
+      this.getMyPledge();
     }
   }
 
@@ -97,8 +128,10 @@ export default class Pledge extends Component {
     this.setState({ SubmitPledgeModalIsOpen: true });
   };
   closeSubmitPledgeModal = () => {
-    this.props.fetchPledgesAction(this.props.eventSlug);
     this.setState({ SubmitPledgeModalIsOpen: false });
+    this.setState({
+      pledgeButtonDisabled: false
+    });
   };
 
   // Functions for Later/Continue after Update Modal
@@ -106,7 +139,6 @@ export default class Pledge extends Component {
     this.setState({ SubmitUpdateModalIsOpen: true });
   };
   closeSubmitUpdateModal = () => {
-    this.props.fetchPledgesAction(this.props.eventSlug);
     this.setState({ SubmitUpdateModalIsOpen: false });
   };
 
@@ -115,6 +147,9 @@ export default class Pledge extends Component {
   }
 
   onFormSubmit() {
+    this.setState({
+      pledgeButtonDisabled: true
+    });
     let value = this.state.value;
     if (value) {
       this.props.postPledge(value);
@@ -134,29 +169,60 @@ export default class Pledge extends Component {
     const data = {
       treeCount: this.state.updatingTreeCount
     };
-    this.props.updatePledge(data, token);
+    this.props.updatePledge(
+      data,
+      {
+        token: token,
+        version: 'v1.3'
+      },
+      this.state.loggedIn
+    );
     this.closeIncreasePledgesModal();
     this.openSubmitUpdateModal();
   };
 
+  getMyPledge = () => {
+    let userPledges =
+      this.props.entities && this.props.entities.eventPledge
+        ? this.props.pledges &&
+          this.props.pledges.allEventPledges &&
+          this.props.pledges.allEventPledges.length > 0
+          ? typeof this.props.entities.eventPledge !== 'undefined'
+            ? ((userPledges = Object.values(this.props.entities.eventPledge)), // convert object to array
+              userPledges.filter(pledge => {
+                return this.props.pledges.allEventPledges.some(f => {
+                  return f.token === pledge.token && f.email === pledge.email;
+                });
+              }))
+            : null
+          : null
+        : null;
+    this.setState({
+      myPledge: userPledges
+    });
+  };
+
   render() {
-    let myPledge = this.props.myPledge;
+    let selectedPledge = {};
+    if (
+      this.props.pledgeEvents &&
+      this.props.pledgeEvents.pledgeEvents.length > 0
+    ) {
+      selectedPledge = this.props.pledgeEvents.pledgeEvents.filter(
+        val => val.slug === this.props.eventSlug
+      )[0];
+    }
+    document.title = getDocumentTitle(selectedPledge.name);
+
+    let myPledge = this.state.myPledge;
     let pledges =
       this.props.pledges && this.props.pledges.total !== undefined
         ? this.props.pledges
         : null;
 
-    if (pledges) {
-      document.title = getDocumentTitle(pledges.name);
-    }
-
-    return this.state.loading ? (
-      <div className="sidenav-wrapper">
-        <LoadingIndicator />
-      </div>
-    ) : (
+    return pledges && pledges.total !== undefined ? (
       <div className="sidenav-wrapper app-container__content--center">
-        <EventHeaderDetails selectedPledge={pledges} />
+        <EventHeaderDetails selectedPledge={selectedPledge} />
         <div className="pledge_content--center">
           <RecentHighestPledges
             pledges={pledges}
@@ -281,7 +347,7 @@ export default class Pledge extends Component {
             isOpen={this.state.AllPledgesModalIsOpen}
             pledges={pledges}
             closeAllPledgesModal={this.closeAllPledgesModal}
-            name={pledges.name}
+            name={selectedPledge.name}
           />
 
           {/* Modal for showing increase pledge form  */}
@@ -315,7 +381,7 @@ export default class Pledge extends Component {
           />
         </div>
       </div>
-    );
+    ) : null;
   }
 }
